refactor: Rename authproxy->oidcproxy.
