--- internal/pkg/service/appsproxy/proxy/apphandler/authproxy/doc.go
@@ -1,3 +1,0 @@
-// Package authproxy provides factory for OAuth2Proxy library.
-// The logic is inserted before the handler from  the "upstream" package.
-package authproxy
