+++ internal/pkg/service/appsproxy/proxy/apphandler/manager.go
@@ -9,7 +9,7 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/dataapps/api"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/dataapps/appconfig"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/dataapps/auth/provider"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/proxy/apphandler/authproxy"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/proxy/apphandler/oidcproxy"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/proxy/apphandler/upstream"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/proxy/pagewriter"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/appsproxy/syncmap"
@@ -24,7 +24,7 @@ type Manager struct {
 	telemetry        telemetry.Telemetry
 	configLoader     *appconfig.Loader
 	upstreamManager  *upstream.Manager
-	authProxyManager *authproxy.Manager
+	oidcProxyManager *oidcproxy.Manager
 	pageWriter       *pagewriter.Writer
 	handlers         *syncmap.SyncMap[api.AppID, appHandlerWrapper]
 }
@@ -39,7 +39,7 @@ type dependencies interface {
 	Telemetry() telemetry.Telemetry
 	PageWriter() *pagewriter.Writer
 	UpstreamManager() *upstream.Manager
-	AuthProxyManager() *authproxy.Manager
+	OidcProxyManager() *oidcproxy.Manager
 	AppConfigLoader() *appconfig.Loader
 }
 
@@ -49,7 +49,7 @@ func NewManager(d dependencies) *Manager {
 		telemetry:        d.Telemetry(),
 		configLoader:     d.AppConfigLoader(),
 		upstreamManager:  d.UpstreamManager(),
-		authProxyManager: d.AuthProxyManager(),
+		oidcProxyManager: d.OidcProxyManager(),
 		pageWriter:       d.PageWriter(),
 		handlers: syncmap.New[api.AppID, appHandlerWrapper](func(api.AppID) *appHandlerWrapper {
 			return &appHandlerWrapper{lock: &sync.Mutex{}}
@@ -87,11 +87,11 @@ func (m *Manager) newHandler(ctx context.Context, app api.AppConfig) http.Handle
 	}
 
 	// Create authentication handlers
-	authHandlers := make(map[provider.ID]*authproxy.Handler, len(app.AuthProviders))
+	authHandlers := make(map[provider.ID]*oidcproxy.Handler, len(app.AuthProviders))
 	for _, auth := range app.AuthProviders {
 		switch p := auth.(type) {
 		case provider.OIDC:
-			authHandlers[auth.ID()] = m.authProxyManager.NewHandler(app, p, appUpstream)
+			authHandlers[auth.ID()] = m.oidcProxyManager.NewHandler(app, p, appUpstream)
 		default:
 			panic("unknown auth provider type")
 		}
