+++ internal/pkg/service/appsproxy/proxy/apphandler/oidcproxy/config.go
@@ -1,4 +1,4 @@
-package authproxy
+package oidcproxy
 
 import (
 	"context"
