+++ internal/pkg/service/appsproxy/proxy/apphandler/oidcproxy/doc.go
@@ -1,0 +1,3 @@
+// Package oidcproxy provides factory for OAuth2Proxy library.
+// The logic is inserted before the handler from  the "upstream" package.
+package oidcproxy
