+++ internal/pkg/service/appsproxy/proxy/apphandler/oidcproxy/logging/writer.go
