+++ internal/pkg/service/appsproxy/proxy/apphandler/oidcproxy/pagewriter.go
@@ -1,4 +1,4 @@
-package authproxy
+package oidcproxy
 
 import (
 	"net/http"
